requirements.txt
Flask==2.3.2
gunicorn==21.2.0
yt-dlp==2024.4.9
openai==1.3.5
whisperx==3.1.1
ffmpeg-python
librosa
pydub
translate
python-dotenv
requests
tqdm
faster-whisper
edge-tts
torch
numpy

